added util  stuff
